added angular client view and /view route

+++ api.py
@@ -1,12 +1,12 @@
 #!./env/bin/python3
 
-from flask import Flask, jsonify
+from flask import Flask, jsonify, send_from_directory, redirect
 from subprocess import call
 from os import chdir, path, getcwd, listdir
 from shutil import rmtree
 
-host = "0.0.0.0"
-port = "8000"
+host = "127.0.0.1"
+port = "8001"
 repo = "git@github.com:bepatient-fr/app-ionic.git"
 pwd = getcwd()
 print(pwd)
@@ -30,22 +30,36 @@ def build_project(branch_name):
     call(["gulp", "project"])
     chdir(pwd)
     return jsonify(message="Branch cloned",
-                   url="http://vm-crashtest.bepatient.mobi/bpapp/%s/bepatient-app/www-dev/" %branch_name)
+                   url="/view/%s" %branch_name)
 
 
 def delete_dir(branch_name):
-    if path.exists("static/%s" %branch_name):
+    if path.exists("static/%s" %branch_name) and branch_name != 'client':
         rmtree("static/%s" %branch_name)
     return jsonify(message='branch removed')
 
 def list_branches():
     branches = {}
     for branch in listdir("static"):
-        branches[branch] = "http://vm-crashtest.bepatient.mobi/bpapp/%s/bepatient-app/www-dev/" %branch
+        if branch != 'client':
+            branches[branch] = "/view/%s" %branch
     return branches
 
 @app.route("/")
-def index():
+def redirect_index():
+    return redirect("/client", code=302)
+
+@app.route("/client")
+def client():
+    return app.send_static_file('client/index.html')
+
+@app.route('/client/<path:path>')
+def send_js(path):
+    print("rendering from client %s" %path)
+    return send_from_directory('static/client', path)
+
+@app.route("/branch", methods=['GET'])
+def get_branches():
     return jsonify(list_branches())
 
 @app.route("/<branch>", methods=['PUT', 'POST'])
@@ -59,7 +73,16 @@ def delete_branch(branch):
 
 @app.route("/<branch>", methods=['GET'])
 def branch(branch):
-    return app.send_static_file('%s/bepatient-app/www-dev/index.html' %branch)
+    return redirect('/view/%s' %branch, code=302)
+
+@app.route("/view/<branch>/", methods=['GET'])
+def view_index(branch):
+    print("VIEW INDEX")
+    return app.send_static_file('%s/bepatient-app/www/index.html' %branch)
+
+@app.route("/view/<branch>/<path:path>", methods=['GET'])
+def view(branch, path):
+    return send_from_directory('static/%s/bepatient-app/www/' %branch, path)
 
 def main():
     app.run(host=host, port=port, threaded=True)
